move-req-state: Move ReqState dataclass to managers/tokenizer_manager_components/request_state.py
ReqState dataclass moves byte-identical to new module. _init_req_state
method stays on TokenizerManager in this commit; its body still
constructs ``ReqState(...)`` which now resolves via the new import.

Per MECH_COMMIT_SPLIT: prep+move for the method itself land in
``move-init-req-prep`` and ``move-init-req-move``.

Refactor chain ID: move-req-state
