+++ python/sglang/srt/managers/tokenizer_manager.py
@@ -15,7 +15,6 @@
 
 import asyncio
 import copy
-import dataclasses
 import json
 import logging
 import os
@@ -78,6 +77,7 @@ from sglang.srt.managers.schedule_batch import MultimodalDataItem
 from sglang.srt.managers.scheduler import is_health_check_generate_req
 from sglang.srt.managers.scheduler_input_blocker import input_blocker_guard_region
 from sglang.srt.managers.tokenizer_control_mixin import TokenizerControlMixin
+from sglang.srt.managers.tokenizer_manager_components.request_state import ReqState
 from sglang.srt.managers.tokenizer_manager_score_mixin import (
     TokenizerManagerScoreMixin,
 )
@@ -129,71 +129,6 @@ _INCREMENTAL_STREAMING_META_INFO_KEYS = (
     "output_top_logprobs",
     "output_token_ids_logprobs",
 )
-
-
-@dataclasses.dataclass
-class ReqState:
-    """Store the state a request."""
-
-    out_list: List[Dict[Any, Any]]
-    finished: bool
-    event: asyncio.Event
-    obj: Union[GenerateReqInput, EmbeddingReqInput]
-
-    # For performance metrics
-    time_stats: APIServerReqTimeStats
-    last_completion_tokens: int = 1
-    ttft_observed: bool = False
-
-    # For streaming output
-    last_output_offset: int = 0
-
-    # Accumulate text lazily so incremental streaming can emit the incoming
-    # delta directly without rebuilding the full output prefix.
-    text: str = ""
-    text_chunks: List[str] = dataclasses.field(default_factory=list)
-
-    def append_text(self, chunk: str):
-        if chunk:
-            self.text_chunks.append(chunk)
-
-    def get_text(self) -> str:
-        if self.text_chunks:
-            self.text += "".join(self.text_chunks)
-            self.text_chunks.clear()
-        return self.text
-
-    def get_crash_dump_output(self) -> Dict[Any, Any]:
-        out = {}
-        if self.text or self.text_chunks:
-            out["text"] = self.get_text()
-        if self.output_ids:
-            out["output_ids"] = self.output_ids.copy()
-        return out
-
-    # For incremental state update.
-    # TODO(lianmin): do not initialize some lists if not needed.
-    output_ids: List[int] = dataclasses.field(default_factory=list)
-    input_token_logprobs_val: List[float] = dataclasses.field(default_factory=list)
-    input_token_logprobs_idx: List[int] = dataclasses.field(default_factory=list)
-    output_token_logprobs_val: List[float] = dataclasses.field(default_factory=list)
-    output_token_logprobs_idx: List[int] = dataclasses.field(default_factory=list)
-    input_top_logprobs_val: List[List[float]] = dataclasses.field(default_factory=list)
-    input_top_logprobs_idx: List[List[int]] = dataclasses.field(default_factory=list)
-    output_top_logprobs_val: List[List[float]] = dataclasses.field(default_factory=list)
-    output_top_logprobs_idx: List[List[int]] = dataclasses.field(default_factory=list)
-    input_token_ids_logprobs_val: List = dataclasses.field(default_factory=list)
-    input_token_ids_logprobs_idx: List = dataclasses.field(default_factory=list)
-    output_token_ids_logprobs_val: List = dataclasses.field(default_factory=list)
-    output_token_ids_logprobs_idx: List = dataclasses.field(default_factory=list)
-
-    # For detokenized logprobs
-    input_token_logprobs: List[Any] = dataclasses.field(default_factory=list)
-    output_token_logprobs: List[Any] = dataclasses.field(default_factory=list)
-    input_top_logprobs: List[Any] = dataclasses.field(default_factory=list)
-    output_top_logprobs: List[Any] = dataclasses.field(default_factory=list)
-    input_token_ids_logprobs: List[Any] = dataclasses.field(default_factory=list)
-    output_token_ids_logprobs: List[Any] = dataclasses.field(default_factory=list)
 
 
 def _slice_streaming_output_meta_info(
